platform complete, 3 modes
Do Times makes a number of moves between the move targets equal to moveTimes. On While goes to index 1 of moveTargets while shouldDo is set to true, and returns to index 0 if it is false. On Stood functions the same as On While, but the bool is set on the player overlapping the collider above the platform.